modularize javascript
