+++ package.json
@@ -14,7 +14,12 @@
     },
     "dependencies": {
         "@tiptap/core": "^2.0.3",
+        "@tiptap/extension-bullet-list": "^2.0.3",
         "@tiptap/extension-image": "^2.0.3",
+        "@tiptap/extension-link": "^2.0.3",
+        "@tiptap/extension-list-item": "^2.0.3",
+        "@tiptap/extension-ordered-list": "^2.0.3",
+        "@tiptap/extension-placeholder": "^2.0.3",
         "@tiptap/pm": "^2.0.3",
         "@tiptap/starter-kit": "^2.0.3"
     }
