+++ resources/js/components/card.js
@@ -1,7 +1,8 @@
-
+import {updateTag, createTag} from './tag.js'
+import {get as getCardContainer} from './container.js'
 
 //Note utility function
-function getNewNoteId(currentNoteData, newNoteData) {
+export function getNewNoteId(currentNoteData, newNoteData) {
     let tempCurrentNoteId = new Array();
     let tempNewNoteId = new Array();
 
@@ -16,7 +17,7 @@ function getNewNoteId(currentNoteData, newNoteData) {
     return tempNewNoteId.filter(newId => !tempCurrentNoteId.includes(newId));
 }
 
-function getRemovedNoteId(currentNoteData, newNoteData) {
+export function getRemovedNoteId(currentNoteData, newNoteData) {
     let tempCurrentNoteId = new Array();
     let tempNewNoteId = new Array();
 
@@ -31,7 +32,7 @@ function getRemovedNoteId(currentNoteData, newNoteData) {
     return tempCurrentNoteId.filter(oldId => !tempNewNoteId.includes(oldId));
 }
 
-function getUpdatedNoteId(currentNoteData, newNoteData) {
+export function getUpdatedNoteId(currentNoteData, newNoteData) {
     let updatedNoteId = new Array();
 
     for (let i = 0; i < currentNoteData.length; i++) {
@@ -49,8 +50,9 @@ function getUpdatedNoteId(currentNoteData, newNoteData) {
 //Note UI manipulation
 
 //newNoteId take one note id
-function addNote(newNoteId, newNotesData) {
+export function addNote(newNoteId, newNotesData) {
     let newCardData = newNotesData.filter(note => note.id == newNoteId);
+    let mediaItemContainer = getCardContainer();
 
     newCardData.forEach(function(card) {
         $(mediaItemContainer).prepend(createCard(card.title, card.body, card.id, card.lastupdated));
@@ -60,8 +62,9 @@ function addNote(newNoteId, newNotesData) {
 }
 
 //Removednotesid take array of note id
-function removeNote(removedNotesId) {
+export function removeNote(removedNotesId) {
     let currentNoteNode = document.querySelectorAll(".note");
+    let mediaItemContainer = getCardContainer();
 
     if (removedNotesId.length < 1) return;
 
@@ -75,7 +78,7 @@ function removeNote(removedNotesId) {
 }
 
 //updatedNoteId take array of update id
-function updateNote(updatedNoteId, newNoteData) {
+export function updateNote(updatedNoteId, newNoteData) {
     let updatedNoteData = Array.from(newNoteData).filter(newNote => updatedNoteId.includes(newNote.id));
     
     updatedNoteData.forEach(noteData => {
@@ -87,16 +90,17 @@ function updateNote(updatedNoteId, newNoteData) {
     return;
 }
 
-function clearCardContainer() {
+export function clearCardContainer() {
     document.querySelector("#card-container").innerHTML = "";
 }
 
-function refreshCardContainerLayout() {
+export function refreshCardContainerLayout() {
+    let mediaItemContainer = getCardContainer();
     $(mediaItemContainer).masonry('reloadItems');
     $(mediaItemContainer).masonry('layout');
 }
 
-function createCard(title, text, id, lastupdated) {
+export function createCard(title, text, id, lastupdated) {
 
     let card = document.createElement("div");
     let innerdiv = document.createElement("div");;
@@ -129,21 +133,21 @@ function createCard(title, text, id, lastupdated) {
 }
 
 //Update card logic
-function updateCardContainerByTag(tagName, clearContainer = false) {
+export function updateCardContainerByTag(tagName, clearContainer = false) {
     if (clearContainer)
         updateCardContainer(tagName, "tag", true);
     else
         updateCardContainer(tagName, "tag", false);
 }
 
-function updateCardContainerBySearch(searchQuery, clearContainer = false) {
+export function updateCardContainerBySearch(searchQuery, clearContainer = false) {
     if (clearContainer)
         updateCardContainer(searchQuery, "search", true);
     else 
         updateCardContainer(searchQuery, "search", false);
 }
 
-function getCurrentCardContainer() {
+export function getCurrentCardContainer() {
     let currentNotesDataHtml = document.getElementsByClassName("note");
     let currentNoteData = new Array();
 
@@ -160,7 +164,7 @@ function getCurrentCardContainer() {
     return currentNoteData;
 }
 
-function updateCardContainer(searchQuery = null, filterType = null, clearContainer = false) {
+export function updateCardContainer(searchQuery = null, filterType = null, clearContainer = false) {
     let currentCard = getCurrentCardContainer();
 
     $.ajax({
@@ -183,11 +187,11 @@ function updateCardContainer(searchQuery = null, filterType = null, clearContain
         }})
 }
 
-function updateCard(currentNoteData, newNoteData) {
+export function updateCard(currentNoteData, newNoteData) {
 
-    newNotesIds = getNewNoteId(currentNoteData, newNoteData);
-    removedNotesId = getRemovedNoteId(currentNoteData, newNoteData);
-    updatedNotesId = getUpdatedNoteId(currentNoteData, newNoteData);
+    let newNotesIds = getNewNoteId(currentNoteData, newNoteData);
+    let removedNotesId = getRemovedNoteId(currentNoteData, newNoteData);
+    let updatedNotesId = getUpdatedNoteId(currentNoteData, newNoteData);
 
     newNotesIds.forEach(newNoteId => addNote(newNoteId, newNoteData));
     removeNote(removedNotesId);
@@ -196,35 +200,35 @@ function updateCard(currentNoteData, newNoteData) {
 
 
 //Event on click note
-$(document).on('click', '.note', function(e) {
+export function initClickNote() {
+    $(document).on('click', '.note', function(e) {
 
-    $('#fullNoteEditor').modal('show');
-    let clickId = $(this).data('id');
-
-    document.querySelector("#fullNoteEditor").setAttribute('data-id', clickId);
-    let tagEditor = document.getElementById("editorTags");
-    let clickedNoteTitle = document.querySelector(`.card.note[data-id="${clickId}"] .card-title`).outerText;
-    let clickedNoteBody = document.querySelector(`.card.note[data-id="${clickId}"] .card-text`).outerText;
-    let clickedNoteTags = document.querySelector(`.card.note[data-id="${clickId}"]`).dataset.tags;
-
-    document.getElementById('titleEditor').value = clickedNoteTitle;
-    document.getElementById('bodyEditor').value = clickedNoteBody;
-
-    // Resize note area
-    $("textarea").trigger("input");
-
-    if (clickedNoteTags === undefined)  {
-        tagEditor.innerHTML = "";
-        return;
-    }
-
-
-
-    clickedNoteTags = String(clickedNoteTags).split(",");
+        $('#fullNoteEditor').modal('show');
+        let clickId = $(this).data('id');
     
-    tagEditor.innerHTML = "";
-    clickedNoteTags.forEach(tag => {
-        tagEditor.append(createTag(tag));
-    });
-
-})
+        document.querySelector("#fullNoteEditor").setAttribute('data-id', clickId);
+        let tagEditor = document.getElementById("editorTags");
+        let clickedNoteTitle = document.querySelector(`.card.note[data-id="${clickId}"] .card-title`).outerText;
+        let clickedNoteBody = document.querySelector(`.card.note[data-id="${clickId}"] .card-text`).outerText;
+        let clickedNoteTags = document.querySelector(`.card.note[data-id="${clickId}"]`).dataset.tags;
+    
+        document.getElementById('titleEditor').value = clickedNoteTitle;
+        document.getElementById('bodyEditor').value = clickedNoteBody;
+    
+        // Resize note area
+        $("textarea").trigger("input");
+    
+        if (clickedNoteTags === undefined)  {
+            tagEditor.innerHTML = "";
+            return;
+        }
+    
+        clickedNoteTags = String(clickedNoteTags).split(",");
+        
+        tagEditor.innerHTML = "";
+        clickedNoteTags.forEach(tag => {
+            tagEditor.append(createTag(tag));
+        });
+    
+    })
+}
