+++ resources/js/init.js
@@ -1,36 +1,47 @@
 
-import initEditor from './editor-tiptap.js'
+import initEditor from './wysiwyg/tiptap.js'
+import {initCardContainer, updatePage} from './components/container.js'
+import {initClickNote} from './components/card.js'
+import { initSidebar } from './components/sidebar.js';
+import { initSearch } from './components/search.js';
+import { initPopover, initPreventDefaultBehavior } from './misc/misc.js';
+import {initClickSimpleEditor, initClickCancelEditor, initClickSubmitEditor, initClickUpdateEditorBtn, initHiddenEditor, initClickEditorTagList, initClickRemoveEditorBtn} from './components/editor.js'
+import { initBlurEditTagInput, initClickAddTagBtn, initClickCloseTagManager, initClickDeleteTagBtn, initClickEditTagBtn, initClickOpenTagManager, initClickSubmitTag } from './components/tag.js';
+
 const editor = initEditor();
 
-var mediaItemContainer = $('#card-container');
-var $mediaItemContainer = mediaItemContainer;
-let currentPageTag = null;
-let currentSearchTerm = null; 
+$('#bulletBtn').on('click', function() {
+    editor.commands.toggleBulletList();
+})
 
-//Init popover for dropdown tag note
-const popoverTriggerList = document.querySelectorAll('[data-bs-toggle="popover"]')
-const popoverList = [...popoverTriggerList].map(popoverTriggerEl => new bootstrap.Popover(popoverTriggerEl))
+initPreventDefaultBehavior();
+initPopover();
 
-//Init masonry
-mediaItemContainer.masonry( {
-    itemSelector: '.card',
-    isFitWidth: true,
-    transitionDuration: '0.4s'
-});
+initClickNote();
+initCardContainer();
+initSidebar();
+initSearch();
 
-initTextAreaToResizable()
-let contentRefresh = window.setInterval(updatePage, 5000);
+initBlurEditTagInput();
+initClickAddTagBtn();
+initClickOpenTagManager();
+initClickCloseTagManager();
+initClickDeleteTagBtn();
+initClickEditTagBtn();
+initClickDeleteTagBtn();
+initClickSubmitTag();
 
+
+initClickSimpleEditor();
+initClickCancelEditor();
+initClickSubmitEditor();
+initClickUpdateEditorBtn();
+initClickRemoveEditorBtn();
+initClickEditorTagList();
+initHiddenEditor();
+
+window.setInterval(updatePage, 5000);
 updatePage();
 
-function updatePage() {
-    if (currentPageTag)
-        updateCardContainerByTag(currentPageTag, false);
-    else if (currentSearchTerm) 
-        updateCardContainerBySearch(currentSearchTerm, false);
-    else 
-        updateCardContainer();
 
-    refreshCardContainerLayout();
-}
