--- resources/js/misc.js
@@ -1,3 +1,0 @@
-$("#tagNoteBtn").on('click', function(e) { e.preventDefault();})
-
-$(".dropdown-menu").on('click', function (e) { e.stopPropagation(); })
