+++ resources/views/home/home.blade.php
@@ -33,6 +33,7 @@
                         <div class='col-sm-3 mx-auto'> 
                             <div id='simpleEditor' class='input-group mb-3'> 
                                 <input type='text' class='form-control shadow-none' placeholder='Take a note..' >
+                                
                                 <!-- <button class='btn btn-light btn-outline-secondary' id='cancel-btn'><i class='bi bi-list-task'></i></button>
                                 <button class='btn btn-light btn-outline-secondary'><i class='bi bi-image'></i></button> -->
                             </div>
@@ -43,11 +44,11 @@
                                         @csrf
                                         <div id='fullEditorCard' class='card-body' >
                                                 <input type='text' name='title' id='titleTextArea' class='form-control border border-0 shadow-none' placeholder='Title'>
-                                                <div id='bodyTextArea' class="element form-control border border-0 shadow-none"></div>
+                                                <div id='fullEditorTextArea' class="element pt-4 form-control border border-0 shadow-none"></div>
                                                 <!-- <textarea name='body' id='bodyTextArea' class='form-control border border-0 shadow-none' placeholder='Take a note..' id='fullEditorTextArea' cols='30' rows='3'></textarea> -->
                                         </div>
                                         <div class='card-footer d-flex flex-row'>
-                                            <button class='btn btn-light btn-outline-secondary border border-0'><i class='bi bi-list-task'></i></button>
+                                            <button id="bulletBtn" class='btn btn-light btn-outline-secondary border border-0'><i class='bi bi-list-task'></i></button>
                                             <button class='btn btn-light btn-outline-secondary border border-0'><i class='bi bi-image'></i></button>
                                             
                                             <button class='btn btn-light border border-0 ms-auto' id='cancelBtn'><i class='bi bi-x'></i></button>
@@ -72,4 +73,17 @@
             </main>
         </div>
     </div>
+    <style>
+        .element p.is-editor-empty:first-child::before {
+            content: attr(data-placeholder);
+            float: left;
+            color: #adb5bd;
+            pointer-events: none;
+            height: 0;
+        }
+
+        .ProseMirror:focus {
+            outline: none;
+        }
+    </style>
 @endsection
