+++ resources/views/layouts/default.blade.php
@@ -5,9 +5,6 @@
 @section('stylesheets')
     <style>
         @include('common.user-stylesheets')
-        .ProseMirror:focus {
-            outline: none;
-            }
     </style>
 
     <link rel="stylesheet" href="{{asset('css/navbar.css')}}">
